requirements.txt
aiofiles==0.6.0
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
fastapi==0.63.0
h11==0.12.0
pycparser==2.20
pydantic==1.7.3
python-telegram-bot==13.1
pytz==2020.5
six==1.15.0
starlette==0.13.6
tinydb==4.3.0
tornado==6.1
tzlocal==2.1
uvicorn==0.13.3
